% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(fc_data, fc_data_ensembled)
}
\arguments{
\item{fc_data}{class: "tbl_df" "tbl" "data.frame" with columns (.mean, PowerConsum, DateIndex).
Use \link{load_all_model_results} to get data from the .rds models.}

\item{fc_data_ensembled}{class: "tbl_df" "tbl" "data.frame" with columns (.mean, PowerConsum, DateIndex, .model).
Use \link{load_ensembled_models} to get data from ensembled .rds models.}
}
\description{
This function will calculate MAPE, MAE and RMSE for all your Forecasts
}
