% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_model_results.R
\name{load_all_model_results}
\alias{load_all_model_results}
\title{Title}
\usage{
load_all_model_results(
  days_to_forecast,
  months_to_forecast,
  year_to_forecast,
  starting_month,
  smard_fc,
  real_data
)
}
\arguments{
\item{days_to_forecast}{Number of Days you want to include in your Forecast. If you need to forecast more then
1 month you could use days_to_forecast > 31.}

\item{months_to_forecast}{Number of months you want to include in your Forecast.}

\item{year_to_forecast}{The year you want to forecast.}

\item{starting_month}{Depend on where you fit ends, you need a "starting_month" Could be 1 for January.}

\item{smard_fc}{SMARD Forecast Data}

\item{real_data}{Cleaned PowerConsum Data
Use \link{load_power_consum()} here to get "smard_fc" and "real_data"}
}
\value{
list(combined_forecasts = combined_forecasts, raw_forecasts = raw_forecasts). Will return a list of combined Forecasts
and Raw Forecasts (nested).
}
\description{
Title
}
\examples{

\dontrun{
fc <- load_all_model_results(
  days_to_forecast = 40,
  months_to_forecast = 6,
  year_to_forecast = 2024,
  starting_month = 1,
  smard_fc = cleaned_smard_pred, # load_power_consum(path)
  real_data = cleaned_power_consum # load_power_consum(path)
)

all_forecasts <- fc$combined_forecasts
raw_fc <- fc$raw_forecasts
}
}
