% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ensembled_models.R
\name{load_ensembled_models}
\alias{load_ensembled_models}
\title{Loading all ensembled Models}
\usage{
load_ensembled_models(
  days_to_forecast,
  months_to_forecast,
  year_to_forecast,
  starting_month,
  real_data,
  smard_fc
)
}
\arguments{
\item{days_to_forecast}{i.r. can be more then 32 will forecast max. of this days}

\item{months_to_forecast}{Will Forecast the number of this months}

\item{year_to_forecast}{Will Forecast this year}

\item{starting_month}{Should be 1 (January) if your Fit ends on 31.12}

\item{real_data}{Check \link{example/forecast.Rmd} should be cleaned_power_consum}

\item{smard_fc}{Data of smard_fc check \link{example/forecast.Rmd} should be cleaned_smard_pred}
}
\description{
This function will load all LHM and DHM models and combine them together, make forecasts.
}
