% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_power_consum}
\alias{load_power_consum}
\title{Will load Power Consum Data and clean up}
\usage{
load_power_consum(path)
}
\arguments{
\item{path}{path Should be the path to your .csv File}
}
\value{
list(raw_data = raw_data, cleaned_data = cleaned_data). Return raw data and cleaned
}
\description{
Will load Power Consum Data and clean up
}
\examples{
\dontrun{
# Define datapaths
power_consum_path <- "dataset\\\\stunde_2015_2024\\\\Realisierter_Stromverbrauch_201501010000_202407090000_Stunde.csv"
power_consum_smard_prediction_path <- "dataset\\\\propgnose_vom_smard\\\\Prognostizierter_Stromverbrauch_202401010000_202407090000_Stunde.csv"

# Load Smard Predictions
power_consum_smard_prediction_loaded <- load_power_consum(path=power_consum_smard_prediction_path)
raw_smard_pred <- power_consum_smard_prediction_loaded$raw_data
cleaned_smard_pred <- power_consum_smard_prediction_loaded$cleaned_data

cleaned_smard_pred <- cleaned_smard_pred |>
  mutate(.model = "SMARD")
names(cleaned_smard_pred)[names(cleaned_smard_pred) == "PowerConsum"] <- ".mean"

# Load PowerConsum Data
power_consum_loaded <- load_power_consum(path=power_consum_path)
raw_power_consum <- power_consum_loaded$raw_data
cleaned_power_consum <- power_consum_loaded$cleaned_data
}
}
