% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_column.R
\name{plot_by_column}
\alias{plot_by_column}
\title{This function will create plots for a selected column.}
\usage{
plot_by_column(df, x, y, title, x_label, y_label, file_name)
}
\arguments{
\item{df}{Your data as DataFrame or Time-Series.}

\item{x}{Select a Column for x-values.}

\item{y}{Select a Column for y-values.}

\item{title}{Title of the Plot (show in the header).}

\item{x_label}{Label of x-axis.}

\item{file_name}{Plot will be saved with this name.}

\item{y_labe}{Label of y-axis.}
}
\value{
None
}
\description{
This function will create plots for a selected column.
}
\examples{
\dontrun{

# Load PowerConsum Data
power_consum_loaded <- load_power_consum(path=power_consum_path)
raw_power_consum <- power_consum_loaded$raw_data
cleaned_power_consum <- power_consum_loaded$cleaned_data

plot_by_column(
  df = cleaned_power_consum,
  x = "Hour",
  y = "PowerConsum",
  x_label = "Stunden",
  y_label = "Stromverbrauch [MW]",
  file_name = "plots\\\\hour_boxplot.png",
  title = "Übersicht der einzelnen Stunden"
)
}

}
