% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_groups.R
\name{plot_by_group}
\alias{plot_by_group}
\title{Creat a faceplot for a group.}
\usage{
plot_by_group(
  filtered_power_consum,
  group_name,
  file_name,
  colors,
  title,
  y,
  x_label,
  y_label
)
}
\arguments{
\item{filtered_power_consum}{time-series.}

\item{group_name}{Character, the column you want to group by.}

\item{file_name}{Character, the name of the file you want to save.}

\item{colors}{Dictionary, define your own color for a specific group.}

\item{title}{Title of the Plot}

\item{y}{Y-Values}

\item{x_label}{Label of X}

\item{y_label}{Label of Y}
}
\description{
This function will create plots for a selected group from a timeseries.
}
