% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calculated_features.R
\name{plot_calculated_features}
\alias{plot_calculated_features}
\title{Will Plot two variables against each other as as geom_point plot.}
\usage{
plot_calculated_features(
  cleaned_power_consum,
  file_name,
  x,
  y,
  x_label,
  y_label
)
}
\arguments{
\item{cleaned_power_consum}{your data}

\item{file_name}{Will be safed here}

\item{x}{X-Values}

\item{y}{Y-Values}

\item{x_label}{Label for x-axis}

\item{y_label}{Label for y-axis}
}
\description{
Will Plot two variables against each other as as geom_point plot.
}
\examples{

\dontrun{
power_consum_path <- "dataset\\\\stunde_2015_2024\\\\Realisierter_Stromverbrauch_201501010000_202407090000_Stunde.csv"
# Load PowerConsum Data
power_consum_loaded <- load_power_consum(path=power_consum_path)
raw_power_consum <- power_consum_loaded$raw_data
cleaned_power_consum <- power_consum_loaded$cleaned_data

cleaned_power_consum$localName[is.na(cleaned_power_consum$localName)] = "Working-Day"

cleaned_power_consum$MeanLastWeek <- rollapply(cleaned_power_consum$PowerConsum, width = 24*8, FUN = function(x) mean(x[1:(24*8-25)]), align = "right", fill = NA) 

cleaned_power_consum$MeanLastTwoDays <- rollapply(cleaned_power_consum$PowerConsum, width = 24*3, FUN = function(x) mean(x[1:(24*3-25)]), align = "right", fill = NA) 

cleaned_power_consum$MaxLastOneDay <- rollapply(cleaned_power_consum$PowerConsum, width = 24*2, FUN = function(x) max(x[1:(24*2-25)]), align = "right", fill = NA) 

cleaned_power_consum$MinLastOneDay <- rollapply(cleaned_power_consum$PowerConsum, width = 24*2, FUN = function(x) min(x[1:(24*2-25)]), align = "right", fill = NA) 

plot_calculated_features(
cleaned_power_consum = cleaned_power_consum,
file_name = "plots/MinLastOneDay.png", 
x = "MinLastOneDay", 
y = "PowerConsum", 
x_label = "Minimaler Stromverbrauch vom letzten Tag [MW]",
y_label = "Stromverbrauch [MW]"
)
}
}
