% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_with_smard.R
\name{plot_compare_with_smard}
\alias{plot_compare_with_smard}
\title{Will compare Residuals for SMARD with LHR + DHR Model, also ACF / PACF Plots for SMARD and LHR + DHR}
\usage{
plot_compare_with_smard(all_forecasts, name_of_best_model)
}
\arguments{
\item{all_forecasts}{Should be all Forecasted values for all Models.}

\item{name_of_best_model}{The name of the best model (could be "version_0" for example).}
}
\description{
Will compare Residuals for SMARD with LHR + DHR Model, also ACF / PACF Plots for SMARD and LHR + DHR
}
\examples{
\dontrun{
ensembled_fc <- load_ensembled_models(
  days_to_forecast = 40,
  months_to_forecast = 6,
  year_to_forecast = 2024,
  starting_month = 1,
  real_data = cleaned_power_consum,
  smard_fc = cleaned_smard_pred,
  model_path = "ensemble_model"
)
all_forecasts_ensembled <- ensembled_fc$all_forecasts
raw_fc_ensembled <- ensembled_fc$raw_forecasts

# Plot best Model for ensembled Models
name_of_best_model_ensembled <- plot_forecast_ensembled(
  all_forecasts = all_forecasts_ensembled,
  metric_results = metric_results,
  cleaned_power_consum = cleaned_power_consum,
  month_to_plot = 1,
  days_to_plot = 40
)

plot_compare_with_smard(
  all_forecasts = all_forecasts_ensembled, 
  name_of_best_model = name_of_best_model_ensembled
)

}


}
