% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_by_group.R
\name{plot_histogram_by_group}
\alias{plot_histogram_by_group}
\title{Will create a Histogram for different Groups.}
\usage{
plot_histogram_by_group(
  filtered_power_consum,
  group_name,
  file_name,
  colors,
  x,
  x_label,
  y_label,
  name_0,
  name_1
)
}
\arguments{
\item{filtered_power_consum}{Time Series you want to Plot.}

\item{group_name}{Column which represents a single group.}

\item{file_name}{Path and Name of the File, where the plot will be safed.}

\item{colors}{Self-Defined Colors.}

\item{x}{x-Values}

\item{x_label}{Label for x-Axis}

\item{y_label}{Label for y-Axis}

\item{name_0}{Names for "0" from group_name}

\item{name_1}{Names for "1" from group_name}
}
\description{
Will create a Histogram for different Groups.
}
\examples{

\dontrun{
power_consum_loaded <- load_power_consum(path=power_consum_path)
raw_power_consum <- power_consum_loaded$raw_data
cleaned_power_consum <- power_consum_loaded$cleaned_data

# Plot
plot_histogram_by_group(
  cleaned_power_consum,
  group_name = "WorkDay",
  file_name = "plots\\\\workday_histogram.png",
  colors = c("1" = "#2E9FDF", "0" = "#FC4E07"),
  x="PowerConsum",
  x_label = "Stromverbrauch [MW]",
  y_label = "Häufigkeit",
  name_0 = "Wochenende oder Feiertage",
  name_1 = "Werktag"
)
}

}
