% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw_data.R
\name{plot_raw_data}
\alias{plot_raw_data}
\title{Will plot Raw Time Series with missing values and duplicated values.}
\usage{
plot_raw_data(df, x, y, color, missing_value, file_name)
}
\arguments{
\item{df}{Time Series or a DataFrame you want to Plot.}

\item{x}{x-Values.}

\item{y}{y-Values.}

\item{color}{Column for duplicated_values (TRUE or FALSE if duplicated)}

\item{missing_value}{Column where values are missing within a Time-Series (TRUE or FALSE if missing)}

\item{file_name}{Where the plot will be safed.}
}
\description{
Will plot Raw Time Series with missing values and duplicated values.
}
\examples{
\dontrun{
plot_raw_data(
df = raw_data,
x = "DateFrom",
y = "PowerConsum",
color = "DuplicatedDate",
missing_value = "MissingValue",
file_name = "plots\\\\raw_data.png"
)
}

}
