% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_representation_of_lhm_dhm_components.R
\name{plot_representation_of_lhm_dhm_components}
\alias{plot_representation_of_lhm_dhm_components}
\title{Will represent LHM and DHR as own components.}
\usage{
plot_representation_of_lhm_dhm_components(
  raw_fc_ensembled,
  path_dhm,
  path_lhm,
  from_month,
  to_month
)
}
\arguments{
\item{raw_fc_ensembled}{Should be Forecast Data with the best Model.}

\item{path_dhm}{Path to the ARIMA model (DHM).}

\item{path_lhm}{Path to the Linear-Model (LHM).}

\item{from_month}{From Which Month the Plot should be.}

\item{to_month}{Until Which Month the Plot should be.}
}
\description{
Will represent LHM and DHR as own components.
}
\examples{

ensembled_fc <- load_ensembled_models(
days_to_forecast = 40,
months_to_forecast = 6,
year_to_forecast = 2024,
starting_month = 1,
real_data = cleaned_power_consum,
smard_fc = cleaned_smard_pred,
model_path = "ensemble_model"
)
all_forecasts_ensembled <- ensembled_fc$all_forecasts
raw_fc_ensembled <- ensembled_fc$raw_forecasts

# LHM DHM representation
plot_representation_of_lhm_dhm_components(path_dhm = "ensemble_model/version_5/arima_2021_2023.rds",
                                          path_lhm = "ensemble_model/version_5/holiday_effect_2021_2023.rds",
                                          from_month = 1,
                                          to_month = 1,
                                          raw_fc_ensembled = raw_fc_ensembled)

}
