% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_models.R
\name{generate_models}
\alias{generate_models}
\title{Will generate any Model (just for more overview)}
\usage{
generate_models(model_name, train_power_consum)
}
\arguments{
\item{model_name}{can be for example "model/arima_0_2021_2023.rds".}

\item{train_power_consum}{TODO:}
}
\description{
Can create a fit
}
