% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ensembled_models.R
\name{load_ensembled_models}
\alias{load_ensembled_models}
\title{Loading all ensembled Models}
\usage{
load_ensembled_models(
  days_to_forecast,
  months_to_forecast,
  year_to_forecast,
  starting_month,
  real_data,
  smard_fc,
  model_path
)
}
\arguments{
\item{days_to_forecast}{Number of Days you want to include in your Forecast. If you need to forecast more then
1 month you could use days_to_forecast > 31.}

\item{months_to_forecast}{Number of months you want to include in your Forecast.}

\item{year_to_forecast}{The year you want to forecast.}

\item{starting_month}{Depend on where you fit ends, you need a "starting_month" Could be 1 for January.}

\item{real_data}{Cleaned PowerConsum Data}

\item{smard_fc}{SMARD Forecast Data}

\item{model_path}{Should be the folder with "Versions"

#' Use \link{load_power_consum()} here to get "smard_fc" and "real_data"}
}
\value{
list(all_forecasts = combined_forecasts, raw_forecasts = raw_forecasts). Will return a list of combined Forecasts
and Raw Forecasts (nested).
}
\description{
This function will load all LHM and DHM models and combine them together, make forecasts.
}
\examples{

\dontrun{
ensembled_fc <- load_ensembled_models(
  days_to_forecast = 40,
  months_to_forecast = 6,
  year_to_forecast = 2024,
  starting_month = 1,
  real_data = cleaned_power_consum,
  smard_fc = cleaned_smard_pred,
  model_path = "ensemble_model"
)
all_forecasts_ensembled <- ensembled_fc$all_forecasts
raw_fc_ensembled <- ensembled_fc$raw_forecasts
}
}
