% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_groups.R
\name{plot_by_group}
\alias{plot_by_group}
\title{Creat a faceplot for a group.}
\usage{
plot_by_group(
  filtered_power_consum,
  group_name,
  file_name,
  colors,
  title,
  y,
  x_label,
  y_label
)
}
\arguments{
\item{filtered_power_consum}{time-series.}

\item{group_name}{Character, the column you want to group by.}

\item{file_name}{Character, the name of the file you want to save.}

\item{colors}{Dictionary, define your own color for a specific group.}

\item{title}{Title of the Plot}

\item{y}{Y-Values}

\item{x_label}{Label of X}

\item{y_label}{Label of Y}
}
\description{
This function will create plots for a selected group from a timeseries.
}
\examples{
\dontrun{

# Load PowerConsum Data
power_consum_loaded <- load_power_consum(path=power_consum_path)
raw_power_consum <- power_consum_loaded$raw_data
cleaned_power_consum <- power_consum_loaded$cleaned_data
week_colors <- c(
"Mo" = palette()[1],
"Di" = palette()[1],
"Mi" = palette()[1],
"Do" = palette()[1],
"Fr" = palette()[1],
"Sa" = palette()[2],
"So" = palette()[2]
)

plot_by_group(
  cleaned_power_consum,
  group_name = "Weekday",
  file_name = "plots\\\\weekday_boxplot.png",
  colors = week_colors,
  title = "Übersicht der einzelnen Wochentage",
  y = "PowerConsum",
  y_label="Stromverbrauch [MW]",
  x_label="Jahre"
)
}

}
